+++ ivy/functional/backends/jax/statistical.py
@@ -1,6 +1,5 @@
 # global
 import jax.numpy as jnp
-import jax.lax as jlax
 from typing import Union, Optional, Sequence
 
 
@@ -216,13 +215,3 @@ def einsum(
     equation: str, *operands: JaxArray, out: Optional[JaxArray] = None
 ) -> JaxArray:
     return jnp.einsum(equation, *operands)
-
-
-def igamma(
-    a: JaxArray,
-    /,
-    *,
-    x: JaxArray,
-    out: Optional[JaxArray] = None,
-) -> JaxArray:
-    return jlax.igamma(a=a, x=x)
