+++ ivy/functional/backends/numpy/experimental/statistical.py
@@ -1,6 +1,6 @@
 from typing import Optional, Union, Tuple, Sequence
 import numpy as np
-
+import math
 import ivy  # noqa
 from ivy.func_wrapper import with_unsupported_dtypes
 from . import backend_version
@@ -382,3 +382,20 @@ def cummin(
     elif reverse:
         x = np.minimum.accumulate(np.flip(x, axis=axis), axis=axis, dtype=dtype)
         return np.flip(x, axis=axis)
+
+
+def igamma(
+    a: np.ndarray,
+    /,
+    *,
+    x: np.ndarray,
+    out: Optional[np.ndarray] = None,
+) -> np.ndarray:
+    def igamma_cal(a, x):
+        t = np.linspace(0, x, 10000, dtype=np.float64)
+        y = np.exp(-t) * (t ** (a - 1))
+        integral = np.trapz(y, t)
+        return np.float32(integral / math.gamma(a))
+
+    igamma_vec = np.vectorize(igamma_cal)
+    return igamma_vec(a, x)
