+++ ivy/functional/backends/numpy/statistical.py
@@ -1,5 +1,4 @@
 # global
-import math
 import numpy as np
 from typing import Union, Optional, Sequence
 
@@ -254,20 +253,3 @@ def einsum(
 
 
 einsum.support_native_out = True
-
-
-def igamma(
-    a: np.ndarray,
-    /,
-    *,
-    x: np.ndarray,
-    out: Optional[np.ndarray] = None,
-) -> np.ndarray:
-    def igamma_cal(a, x):
-        t = np.linspace(0, x, 10000, dtype=np.float64)
-        y = np.exp(-t) * (t ** (a - 1))
-        integral = np.trapz(y, t)
-        return np.float32(integral / math.gamma(a))
-
-    igamma_vec = np.vectorize(igamma_cal)
-    return igamma_vec(a, x)
